# openapi.yaml
openapi: 3.0.3
info:
  title: Aetheria API
  description: |
    REST API for the Aetheria shared-world mobile MMORPG.
    All authenticated endpoints expect `Authorization: Bearer <accessToken>`.
    All state-mutating endpoints are server-authoritative.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Local dev
  - url: https://api.aetheria.example
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /players/me:
    get:
      tags: [players]
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Player' }
    patch:
      tags: [players]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                avatarId: { type: string }
                bio: { type: string, maxLength: 280 }
      responses: { '200': { description: Updated } }
  /players/{displayName}:
    get:
      tags: [players]
      security: []
      parameters:
        - in: path
          name: displayName
          required: true
          schema: { type: string }
      responses:
        '200': { description: Public profile }
...
  /characters/{characterId}/quests/available:
    get:
      tags: [quests]
      parameters: [{ in: path, name: characterId, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: OK } }
  /characters/{characterId}/quests/active:
    get:
      tags: [quests]
      parameters: [{ in: path, name: characterId, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: OK } }
  /characters/{characterId}/quests/accept:
    post:
      tags: [quests]
      parameters: [{ in: path, name: characterId, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [questId]
              properties: { questId: { type: string } }
      responses: { '201': { description: Accepted } }
  /characters/{characterId}/quests/abandon:
    post:
      tags: [quests]
      parameters: [{ in: path, name: characterId, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [questId]
              properties: { questId: { type: string } }
      responses: { '200': { description: Abandoned } }
  /characters/{characterId}/quests/turn-in:
    post:
      tags: [quests]
      parameters: [{ in: path, name: characterId, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [questId]
              properties: { questId: { type: string } }
      responses:
        '200':
          description: Rewards granted
          content:
            application/json:
              schema:
                type: object
                properties:
                  rewards:
                    type: object
                    properties:
                      gold: { type: integer }
                      xp: { type: integer }
                      items:
                        type: array
                        items: { type: object, properties: { item_id: { type: string }, qty: { type: integer } } }
...
  /marketplace:
    get:
      tags: [marketplace]
      parameters:
        - { in: query, name: itemId, schema: { type: string } }
        - { in: query, name: maxPrice, schema: { type: integer } }
      responses: { '200': { description: Active listings } }
...
  /marketplace/list:
    post:
      tags: [marketplace]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [inventoryItemId, quantity, priceGold]
              properties:
                inventoryItemId: { type: string, format: uuid }
                quantity: { type: integer, minimum: 1 }
                priceGold: { type: integer, minimum: 1 }
      responses: { '201': { description: Listed } }
  /marketplace/{id}/buy:
    post:
      tags: [marketplace]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Purchased } }
  /marketplace/{id}/cancel:
    post:
      tags: [marketplace]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Cancelled } }
...
  /notifications:
    get:
      tags: [notifications]
      parameters: [{ in: query, name: unreadOnly, schema: { type: boolean } }]
      responses: { '200': { description: Notifications } }
  /notifications/{id}/ack:
    post:
      tags: [notifications]
      parameters: [{ in: path, name: id, required: true, schema: { type: string, format: uuid } }]
      responses: { '200': { description: Acked } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    Player:
      type: object
      properties:
        id: { type: string, format: uuid }
        displayName: { type: string }
        avatarId: { type: string, nullable: true }
        bio: { type: string, nullable: true }
        gold: { type: integer }
        gems: { type: integer }
        lastSeenAt: { type: string, format: date-time }
...